Compute BIC score

compute BIC score when fitting a DCSBM to network data

eval_dcsbm_bic(A, z, K, poi)

Arguments

A

adjacency matrix

z

label vector

K

number of community in z

poi

whether to use Poisson version of likelihood

Details

the BIC score is calculated by -2*log likelihood minus \(K\times(K + 1)\times log(n)\)

References

BIC score is originally proposed in Likelihood-based model selection for stochastic block models Wang, YX Rachel, Peter J. Bickel, The Annals of Statistics 45, no. 2 (2017): 500-528.

The details of modified implementation can be found in Adjusted chi-square test for degree-corrected block models, Linfan Zhang, Arash A. Amini, arXiv preprint arXiv:2012.15047, 2020.

See also