Plot a network

Plot a network using degree-modulated node sizes, community colors and other enhancements

plot_net(
  gr,
  community = NULL,
  color_map = NULL,
  extract_lcc = TRUE,
  heavy_edge_deg_perc = 0.97,
  coord = NULL,
  vsize_func = function(deg) log(deg + 3) * 1,
  vertex_border = F,
  niter = 1000,
  vertex_alpha = 0.4,
  remove_loops = T,
  make_simple = F,
  ...
)

Arguments

gr

the network as an igraph object

community

community assignment; vector of node labels

color_map

color palette for clusters in 'gr'

extract_lcc

Extract largest connected component or not

heavy_edge_deg_perc

Degree percentile threshold for determining heavy edges

coord

Optional starting positions for the vertices. If this argument is not NULL then it should be an appropriate matrix of starting coordinates.

vsize_func

function to determine the size of node size

vertex_border

whether to show the border of vertex or not

niter

number of iteration for FR layout computation

vertex_alpha

factor modifying the opacity alpha of vertex; typically in [0,1]

remove_loops

whether to remove loops in the network

make_simple

whether to simplify edge weight calculation

...

other settings