Adjusted spectral test

Adjusted spectral test

adj_spec_test(A, K, z = NULL, DC = T, theta = NULL, B = NULL, ...)

Arguments

A
adjacency matrix

K
number of communities

z
label vector for rows of adjacency matrix. If not given, will be calculated by the spectral clustering

DC
whether or not include degree correction in the parameter estimation

theta
give the propensity parameter directly

B
give the connectivity matrix directly

...
parameters for spectral clustering.

Value

A list of result

stat

NAC or NAC+ test statistic.

z

row label vector.

y

column label vector.